moving yarn / npm install up
Because otherwise I see issues in VS code before even running for the first time as some dependencies are not installed. Thus, I think running yarn should be part of the minimal setup.

That made me realize that we should consider also catering for NPM. It's still the default and some people might just be turned down by not understanding what yarn is and how to use it / where to get it from.